[devtools] fork devtools-indicator (#80456)
Unlike the menu popover (legacy), the panel UI can be decoupled from the
indicator and communicate with the dispatcher. We don't need those refs,
animation, etc. anymore, so removed them as well.

Also, it feels good to move out of the `errors/` directory as it's no
longer confined to the error overlay.