chore(turbopack): Dedupe rustls dependencies by updating reqwest (#80632)
We were pinning the `reqwest` version because of build issues: https://github.com/vercel/next.js/pull/61076 . I don't imagine that's an issue anymore, as we've been shipping the newer version of `reqwest` as part of `wasmer`.

It seems like the tls dependencies can be a little slow to build, so deduping them is nice.

build-and-deploy job: https://github.com/vercel/next.js/actions/runs/15719542395