Respond with `404` for unknown server actions (#80613)
When a server action is sent from an older deployment and skew
protection is not enabled or implemented, we intended to return a `404`
response.

However, this behavior was never properly tested and was accidentally
broken in #48626 when `actionAsyncStorage` was added to the action
handler without a `return`, preventing the early `'not-found'` returns
from taking effect.

This PR restores the `404` response handling and adds end-to-end tests
for all server action variations. It also changes logs from errors to
warnings to reflect that this scenario is expected to occur
occasionally. Warnings remain useful (instead of silent failures), as
users may want to tune their skew protection settings based on how often
they see them.

fixes #75541