Apply normalizers for init URL used for handlers (#80694)
This applies our normalizers to the `initURL` we pass to the handlers
interface as currently it is created before our existing normalizing
takes place which causes us to fail to generate the correct cache keys
for ISR paths. This specifically breaks our de-duping for client segment
cache as it relies on the cache key being consistent across all segment
paths being requested during a revalidation.

In the future we will have to hoist the URL normalizing anyways so this
leaves it as is until then.

This also adds a regression test for the cache key not being normalized
causing extra render passes for client segment cache entries when it
shouldn't have.