omit searchParam data from FlightRouterState before transport (#80734)
Certain types of data gets stashed in the client-side
`FlightRouterState`, such as the "refresh" marker alongside a URL that
the client can callback to when it needs to revalidate segments that
were "activated" on previously visited URLs. Additionally, query
parameter data gets encoded in the `__PAGE__` segment key, but isn't
actually needed by the server: `__PAGE__` segments would always be the
leaf, so it wouldn't cause any unexpected behavior in the diffing logic.

With large query params, this can lead to large amount of data needing
to be sent to the server, only for it to not actually be used. There's a
larger refactor around the server/client transport being handled in the
`experimental.clientSegmentCache` flag, so this is a stop-gap until that
behavior is stabilized.

Closes NEXT-4521