devtools: couple restart dev server UI with persistent cache flag (#80751)
The goal of this UI is to surface a nudge when we think that the server
might be in a bad state due to persistent caching problems. It probably
doesn't make sense to show up in any other case. While we work to figure
out the best way to nudge it in dev, this couples the behavior behind
the persistent caching flag.

---------

Co-authored-by: Sebastian "Sebbie" Silbermann <sebastian.silbermann@vercel.com>