Revert "Respond with `404` for unknown server actions (#80613)" (#80662)
@unstubbable and i independently worked on fixing the same bug, but my
solution got a bit further along + i have a stack of other changes on
top, so we're going to do mine instead.
- #77012 implements equivalent functionality for fetch actions (but more
efficiently, because it skips rendering the 404 HTML, and with a better
error message)
- #77014 also fixes the error message for MPA actions, but unlike the
reverted PR, it _doesn't_ make them return a 404 html page -- that's not
semantically correct, because an unrecognized action error is fixed by a
refresh.
- #78933 allows handling this error on the client