Turbopack Build: Fix metadata-edge test (#80842)
This test is checking if `@vercel/og` gets shared between two routes, it
assumed that `next/og` itself (the wrapper code) would also be chunked,
but the size limit for Turbopack is slightly larger so it doesn't end up
bundling it into a separate chunk. Changed the test to check that
`@vercel/og` is correctly chunked.