Allow intercepting dynamic routes to be partially prerendered (#80851)
For intercepting dynamic routes we do prerender `.prefetch.rsc`, as well
as `.segment.rsc` files if `clientSegmentCache` is enabled. However, we
omitted adding those routes to the prerender manifest, which resulted in
server errors when looking up the RSC files.

I've added two tests to verify the one-line fix, one in
`test/e2e/app-dir/parallel-routes-and-interception` (relevant when tests
are run with `__NEXT_EXPERIMENTAL_PPR=true`), and one in
`test/e2e/app-dir/segment-cache/basic` (has `dynamicIO` and
`clientSegmentCache` enabled). Both tests are written in the style of
the existing tests in the respective test suites.

---------

Co-authored-by: Jimmy Lai <laijimmy0@gmail.com>