Turbopack: fix chunking context caching (#80862)
### What?

Fixes caching by changing arguments from `Option<RcStr>` to `Vc<Option<RcStr>>`.

This changes the argument from passed by value to passed by reference. This makes a big differences regarding caching when the string value is changing.

`Option<RcStr>`: When the string changes a completely new task is created resulting in ChunkingContext being in a new cell, which ends up creating new tasks for all functions that receive the ChunkingContext. This results in a lot of new tasks that need to be computed and stored.

`Vc<Option<RcStr>>`: When the string changes the existing task is invalidated and recomputed. The result is still the same cell and no new tasks for functions that receive the ChunkingContext are created. Only tasks that use the string are invalidated and recomputed.