refactor(turbo-tasks-backend): Use ringmap crate instead of custom DequeSet type for aggregation updates (#80867)
`ringmap` is a pretty new library from the same folks behind `indexmap` (and it's actually a fork of `indexmap`) that uses a `VecDeque` instead of a `Vec`.

https://github.com/indexmap-rs/ringmap/

This adds an extra dependency, but it's a drop-in replacement, and our current `DequeSet` type is pretty far from ideal.

## Benchmarks

I thought this might improve perf, but this looks neutral on codspeed. So it's just a matter of if we think this improves the code or not.