[tests]: add failing deploy test to manifest (#80973)
For the same reason we have the other `unstable_expire<x>` tests
disabled, this adds another one that has been frequently failing for us.
([Linear x-ref](https://linear.app/vercel/issue/ENET-1729))

`app-action-node-middleware` is also a duplicate of `app-action`, but
when we created it, we did not copy over the manifest exclude list.