Turbopack: fix exposed modules in scope hoisting (#81073)
Closes PACK-4943
Closes https://github.com/vercel/next.js/issues/80998

If we have to split up a merged group because multiple chunks have different execution order, we need to make sure that the references that cross merged groups because of this actually work at runtime (i.e. the referenced modules need to expose themselves in the runtime module registry).

Open for ideas on how to optimize this computation. It's the usual question of HashSet vs iterating a list multiple times.

The core problem is: we need to expose all modules that appear in `intra_group_references` as values with a key that is in a different list (i.e. key and value musn't both be in `after`, `common`, or `before`)