[devtools]: move RestartServerButton to panel footer (#81082)
This moves the restart button to the footer in the new panel UI. It also
properly gates to only be enabled with Turbopack + Persistent Cache.

Because it's no longer part of the error overlay, I refactored the logic
a bit to hoist the flag into overlay state and setup the beforeunload
listener from within the error renderer.

This heuristic still feels a bit unreliable and I'm not entirely
convinced how useful it is yet, but the goal with this PR is just to get
it to render in the right spot for now.

![CleanShot 2025-06-30 at 10 24
12@2x](https://github.com/user-attachments/assets/7c96b7ad-0b90-4a08-9b9d-5bea02348ab8)

Closes NEXT-4561