Docs: paper cuts July 1st  (#81109)
Just a couple of adjustments to Caching docs claim that React State is
preserved upon SPA transitions - as of today, without the Activity API,
only shared layouts state is preserved.

And we are saying redirect should be called outside try/catch blocks,
but rather, it is the try block that's the problem. So I am rephrasing
that.

I still think it is best to keep the redirect away from the try/catch
statement altogether.