refactor(turbopack): Implement `BackingStorage` for `Either`, replace `NextTurboTasks` with a type alias (#80823)
The `NextTurboTasks` enum existed to allow us to enable or disable persistent caching at runtime.

This PR implements `BackingStorage` for `Either`, so that the complexity of this branching is moved into `turbo-tasks-backend`.

This eliminates the need for the next.js and eventually other consumers (`turbopack-cli`, Utoo) to handle this branching if they want to support multiple backing storage layers.

The tradeoff is that moving these branches further down in the callstack might lead to more (easily predicted) branches in the code. We can see if codspeed notices anything. **Update:** Codspeed is performance-neutral.