Fix resolvedPathname for middleware rewrite (#81144)
This updates the handling of `resolvedPathname` to handle the case we
don't know the `rewroteURL` for example when it's a middleware rewrite
and we can't derive it since it's fully dynamic. Since we always know
the `srcPage` we can always recreate the final URL using the params
which this does as it's more reliable than relying on the parsed URL as
a backup.

Validated against our deploy tests
https://github.com/vercel/next.js/actions/runs/16014327609/job/45178072108
and
https://github.com/vercel/vercel/actions/runs/16014347060/job/45178081995?pr=13509

x-ref: [slack thread](https://vercel.slack.com/archives/C093RLQ7RGC)