[node-webstreams] Exhaustive React aliases in App Router (#81040)
The previous logic relied heavily on default-fallthroughs which makes it hard to follow what alias we end up using in which bundle.

The new logic has more duplication which is intended so that you think about each entrypoint. With the new factoring, some bad combinations are more obvious (e.g. `react-dom/server` in product code not using the vendored version or `react-server-dom-webpack/server` not throwing in a Client environment). Compare changing aliases to .node in https://github.com/vercel/next.js/pull/81048/ (based on this refactor) vs changing it without the refactor: https://github.com/vercel/next.js/pull/80941

Fixing the bad combinations is not part of this work. We probably need to finish `react-markup` first before we get more strict with `react-dom/server` usage in Server code.