[sourcemaps] ignore-list Webpack runtime (#81229)
The way this works in dev is by assigning a dummy sourcemap to the Webpack runtime that ignore-lists everything. The mappings will will be incorrect in development. Production sourcemaps will be correct since those emit real sourcemaps for the webpack runtime where we just start ignore-listing modules that look like Webpack runtime.

Chrome already has this heuristic where it ignore-lists a source if its sourcemap ignore-lists every source. We'll use the same heuristic in the terminal and Redbox in a follow-up. We'll implement this heuristic for the Redbox and terminal in a follow-up.


https://github.com/user-attachments/assets/b250e61d-c6f9-49b6-85c4-8bdd54275a01

Closes https://linear.app/vercel/issue/NEXT-4409/