Allow `pnpm build` to be run when offline (#81273)
When running `pnpm build` in the Next.js repo without an internet
connection, this currently fails while trying to download the AMP
`validator_wasm.js` file. We can fix this, and thus enable offline
support for `pnpm build`, by moving the download of this file from the
`precompile` task to the `ncc_amphtml_validator` task. The file is then
committed to `src/compiled/amphtml-validator`, and later copied into
`dist/compiled/amphtml-validator` as part of the build command, without
the need to trigger a request.