Remove incorrect PostCSS config type (#81338)
### What?

Currently the docs is incorrectly suggesting to type the PostCSS config
with types from Tailwind CSS. This pull request removed the wrong types.

If desired one can correctly type the config with the
[`postcss-load-config`](https://github.com/postcss/postcss-load-config/blob/main/src/index.d.ts#L52-L60)
package:

```js
/** @type {import('postcss-load-config').Config} */
const config = {
    plugins: {
        '@tailwindcss/postcss': {},
    },
}

export default config
```

However if your package manager does not hoist the package, you need to
install it separately just for the types, which does not seem to be
worth it.

### Why?

### How?

Co-authored-by: JJ Kasper <jj@jjsweb.site>