[use-cache] track cache read earlier when encrypting bound args (#81427)
bound args need to be encrypted for use cache functions and currently we
track the read slightly later than will be necessary when we sync the
latest React. This change moves the tracking to be as early as possible
(as soon as the input signal is aborted or when the bound args are
finished being serialized, whichever is first).